Updates to workflows
Made "skip cleanup" the default operating mode for all workflows. Added a --cleanup flag to force cleanup.
Modified all create methods so that an error does not occur if the object already exists - it is simple returned with a message to the user.
Added a workflow to list objects on the array.
Renamed workflow initiator_group_lifecycle.py to create_and_update_initiator_group.py
Cleaned up and included the nimble_sdk_helper.

Signed-off-by: Brian Sorge <brian.sorge@hpe.com>